latest progress

+++ CSharp-OOP/01.Defining-Classes-1/Startup/Call.cs
@@ -12,10 +12,10 @@
         private string date;
         private string time;
         private int calledNumber;
-        private int callDuration;
+        private double callDuration;
 
         //constructors
-        public Call(string date, string time, int CalledNumber, int CallDuration)
+        public Call(string date, string time, int CalledNumber, double CallDuration)
         {
             this.Date = date;
             this.Time = time;
@@ -27,6 +27,6 @@
         public string Date { get; set; }
         public string Time { get; set; }
         public int CalledNumber { get; set; }
-        public int CallDuration { get; set; }
+        public double CallDuration { get; set; }
     }
 }
